requirements.txt
# ============================================================================
# Telegram AI Bot - Dependencies
# ============================================================================

# Telegram Bot Framework
python-telegram-bot>=21.0

# Google Gemini API (для prompt optimization и voice transcription)
google-generativeai>=0.3.0

# HTTP клиенты (для Webhook сервера и API вызовов)
aiohttp>=3.9.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Replicate API Client (for Z-Image/FLUX)
replicate>=0.25.0

# ============================================================================
# Production (Render) - опционально для продвинутого деплоя
# ============================================================================
# gunicorn>=21.0.0  # Если нужен WSGI сервер (aiohttp достаточно для нашего случая)
